Use a hard-coded branch name when enabling CodeQL
I have struggled with this change, and I am still not sure if it's for the best. This change increases determinism, which improves simplicity, but at the cost of potentially more errors, or rather, less flexibility in the face of such error. The error being: if you make a mistake while passing this, should this tool help you fix it or not handle that case at all, as it _is_ out of scope? I'm leaning toward the latter.