Re-centralize the code into a single file.
This undoes some of my earlier work to fracture the codebase logically. That effort, while well-intentioned, was ultimately a bit silly and made for an odd separation of code. Especially such a small codebase. So, I think I'll just keep things more centrally located in a single file unless/until it gets "too big" for a single file.