Eliminate dependency on the git executable.
This commit marks the --git option as deprecated (and soon to be unsupported), because it is no longer necessary. With the next commit, we will be able to achieve the same functionality, but without shelling out to `git`, simply by extending our use of the GitHub API.

Credit to @ssulei7 for the idea!