Use a different branch name when enabling CodeQL
The old branch name was based on the name of the repository. The idea was that it would be unique enough that the chance of it existing would be nil, and it identifies the software that generated the PR.

However, there was a customer that did not appreciate having my name in there. I think they thought it would be confusing to their developers. I understand. So, this new branch name is generic, but hopefully we don't start running into name conflicts.