Fix "duplicate language matrix key" bug.
When two languages in the repository's set map to the same CodeQL keyword, the list of languages ends with a redundant pair. For example, if the repository contains 'C' and 'C++', both will map to 'cpp', resulting in a build-matrix of ['cpp', 'cpp']. This has annoying consequences, like double the Actions usage. Potentially, even, duplicate alerts? Fortunately, the fix is quite simple: run the list through a `set` first.